Add QCGridPattern brush
Add new brush type for grid patterns. This is modern version of QBrush
cross patterns, which are textures that can be colorized. With
QCGridPattern the grid is done with a shader and colors, line width,
feather and rotation can be adjusted.

Contains example in gallery, basic auto tests and API docs.

Task-number: QTBUG-142049
Change-Id: I97cd49a73971da3aa051a5fb3d72f16a65d6af39
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>