CMake: Fix running of qcshadergen on Windows
Use _qt_internal_get_tool_wrapper_script_path to prepend a batch
script that setups the PATH env to point to the Qt bin dir, so that
Qt DLL dependencies are found on Windows hosts.

Wrap the qcshadergen executable with TARGET_FILE to ensure the path
to the tool is computed correctly in various configurations, like
per-repo / top-level / standalone.

Wrap the qcshadergen executable with COMMAND_CONFIG when using
Ninja multi-config generators and CMake >= 3.20, to ensure we only
build the release version of the tool by default.

Change-Id: I9c2101d0f54296d39f353886054e1f755e62ad2c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>