Enable qcrhiplumbing autotest in CI, with fixes
Make QCOffscreenCanvas more robust. Avoid storing the (explicitly
shared) objects themselves in the renderer. Rather, store the internal
struct with the QRhi resources. Avoid the mess iterating through the
list (while potentially modifying it) creates in renderDelete() by
swapping with a new temporary first.

We also raise the minimum GLSL version for OpenGL to 130,
meaning OpenGL 3.0. This is due to textureSize() which is not
in GLSL 120.

For macOS and OpenGL, the expectation from now on is that a 3.2
or 4.1 core profile context is requested by the application,
the default 2.1 won't do. On other platforms that have proper
OpenGL versioning this is a non-issue. In practice it won't
matter much for macOS either, since we expect Metal to be used.

Change-Id: Ib9052d10d94124a4d62aec9a4f7cf4c82ad5c329
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>