Combine reserve methods in QCPainterPath
Combine these reserve methods into a single one. There is really never
a need to call just one of them, if the amount of commands changes so
does data points and vice versa. This makes it also to match with the
constructor version, and we can do default data points reserving.

Change-Id: I340edf0cb72149ae6b76445c5fb5e88bb8f3e416
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>