Add clearRect
The implementation is basically fillRect() with a custom color and
having blending disabled, because things like the clip rect are
expected to be supported by clearRect() too.

The savings, with regards to performance, are expected to come
from having blending disabled, thus reducing the pixel processing
load on the GPU, especially when the rect is large.

Other than that, clearRect() is identical to fillRect(). Except of
course that fillRect() cannot achieve the blending-less writing out
of alpha == 0 pixels (since it would always blend).

Exercise it in gallery, with clipping even.

Fixes: QTBUG-142430
Change-Id: Ifa98da73b4b76eb2e93b3c57d917e7d093385476
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>