Cleanup initializeResources impl in item
Rename the variable that is used for nothing but
to keep track if initializeResources() was called.

It could be argued that initializeResources() should be
called after the first synchronize(), to allow using
data copied from the main thread, but it turns out
QQuickRhiItem also calls its initialize() before
synchronize().

Then again, the semantics of that initialize() and our
initializeResources() are quite different in practice,
f.ex. when it comes when these methods are called.

Still, keep it as-is for now. (invoke init before sync)
Add a highlight in the docs.

Pick-to: 6.11
Change-Id: I8aea5d654f84f712520e292570b04a7879858d3b
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>