Make QCGallery scalable
Make QCGallery to scale different aspect ratios and portrate/landscape.
This is done so that multiple views can be viewed at the same time,
in landscape mode ~3 and in portrate ~1. TopBar content is shown for
the currently center view and the views (partly) visible on sides have
decreased opacity and saturation.

Noticed that resizing the Painter Paths view doesn't work correctly.
To be investigated.

Pick-to: 6.11
Change-Id: I480d261d1a1f62df4411caee13636511b9b646a6
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>