Use smoothstep in shape antialiasing
Previously antialiasing was applied linearly, which looks OK when
antialiasing a single pixel. But as we allow freely adjusting the
antialiasing amount, apply it with smoothstep instead. This way then
antialiased paths are used as a shadow/glow, the appearance is
closer to blur effect. Text antialiasing is already using smoothstep.

Pick-to: 6.11
Change-Id: If2722e33fe6f137298a6c7a585355b451ada9d5c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>